Revert "split test configuration out"
This reverts commit b9623fa212662bf0889841bf9afb370da09bd94f.
