+++ api/src/integration/java/org/pmiops/workbench/BaseIntegrationTest.java
@@ -23,6 +23,18 @@ public abstract class BaseIntegrationTest {
 
   protected static WorkbenchConfig config;
 
+  @TestConfiguration
+  static class Configuration {
+    // This prototype-scoped bean override will cause all autowired services to call this method
+    // for their Provider<WorkbenchConfig>. Further modifications may be made from within each test
+    // case.
+    @Bean
+    @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
+    WorkbenchConfig getIntegrationTestConfig() throws IOException {
+      return config;
+    }
+  }
+
   @Before
   public void setUp() throws IOException {
     config = loadTestConfig();
