+++ api/src/test/java/org/pmiops/workbench/api/WorkspacesControllerTest.java
@@ -19,8 +19,6 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyZeroInteractions;
 import static org.mockito.Mockito.when;
-import static org.pmiops.workbench.BaseTestConfiguration.CLOCK;
-import static org.pmiops.workbench.BaseTestConfiguration.NOW;
 import static org.pmiops.workbench.utils.TestMockFactory.DEFAULT_GOOGLE_PROJECT;
 
 import com.google.api.services.cloudbilling.Cloudbilling;
